Move ops utils to internal/ops/lifecycle
